+++ lib/osm/event.rb
@@ -35,7 +35,7 @@ module Osm
 
     attr_accessible :id, :section_id, :name, :start, :finish, :cost, :location, :notes, :archived
 
-    validates_numericality_of :id, :only_integer=>true, :greater_than=>0
+    validates_numericality_of :id, :only_integer=>true, :greater_than=>0, :allow_nil => true
     validates_numericality_of :section_id, :only_integer=>true, :greater_than=>0
     validates_presence_of :name